+++ Changelog.md
@@ -1,5 +1,11 @@
 ###
 
+### 05/02/2022
+ - Added forward pass of neural network to controll car
+ - Added fitness evalutation (Average Speed i.e (ditance gates hit / time taken))
+ - Changed to use JSON to store load and work with the car weights and biases
+ - Changed the weight and bias generation to allow values between -1 and 1
+
 ### 21/11/2021
  - Added networks and network handler
  - Send acceleration and turn values to Unity from the NN scripts